[#2] Refactor: Order Table에서 Table Head와 Table Body로 분리 및 OrderTable폴더 생성
head와 body로 컴포넌트를 분리했고 OrderTable폴더를 만들어 각 컴포넌트를 넣었습니다.
